Raise errors when population is zero
This adds two checks for the situation that caused Milton Keynes to fail
during final import and bring the API down with it.  The root of the problem
was that there was no population in the imported blocks within the neighborhood
boundary, but the way that error manifested was not helpful. It ran the whole
analysis and imported the broken score results, then failed because the blank
population value caused an error on the database level for any query that
tried to load the record.

This adds two pieces of error-checking:
- one early in the analysis, as soon as the blocks are imported, so that it will
raise a more informative error and avoid running a whole analysis with bad data
- a second in the command that imports analysis results into the database, to
raise an error if it encounters an empty population value, rather than importing
it and causing the database to start throwing errors.

Resolves #911

+++ src/analysis/import/import_neighborhood.sh
@@ -190,6 +190,14 @@ then
         echo "Census Blocks in analysis: ${BLOCK_COUNT}"
         set_job_attr "census_block_count" "${BLOCK_COUNT}"
 
+        POPULATION=$(psql -h "${NB_POSTGRESQL_HOST}" -U "${NB_POSTGRESQL_USER}" -d "${NB_POSTGRESQL_DB}" \
+            -t -c "SELECT SUM(pop10) FROM neighborhood_census_blocks;")
+        echo "Population is ${POPULATION}"
+        if [[ $POPULATION -eq 0 ]]; then
+            echo "Error: Total population of included census blocks is zero."
+            exit 1
+        fi
+
         # Remove NB_TEMPDIR
         rm -rf "${NB_TEMPDIR}"
     fi
